fix(stream): rename validateStream to validateStages
There was a semantic miss between createStream and validateStream. One creates a “stream” from stages while the other was validating that a stream can be created from stages. Semantically  the output from createStream should have been compatible with validateStream but that wasn’t the case. validateStages isn’t as attractive of a name but it is more accurate.